+++ CHANGES.rst
@@ -31,7 +31,7 @@ Changelog
     VOEvents in the playground instance. Fix this by having
     ``gwcelery.tasks.gcn.validate`` discard all VOEvents if the VOEvent
     broadcaster is disabled.
-    
+
 -   Update ligo-followup-advocate (the circular generating repo) version to
     0.0.27.