+++ CHANGES.rst
@@ -32,8 +32,7 @@ Changelog
     ``gwcelery.tasks.gcn.validate`` discard all VOEvents if the VOEvent
     broadcaster is disabled.
 
--   Update ligo-followup-advocate (the circular generating repo) version to
-    0.0.27.
+-   Update ligo-followup-advocate to 0.0.27.
 
 0.4.3 (2019-04-05)
 ------------------